+++ Changelog.md
@@ -18,3 +18,5 @@
   * TypeCheck has new method: IsNumerical(t)
   * TypeExtensions has new method: GetBaseTypeHierarchy(t)
   * TypeCheck has new method: IsNewable(t)
+  * Infinite recursions for self referencing types has been avoided in ObjectInstantiator
+  * Infinite recursions for __TreatAsLeaf__ nodes would be avoided.